pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kb-arena"
version = "0.1.0"
description = "Benchmark retrieval strategies on your documentation — find which KB architecture fits your data"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Xavier Puspus", email = "xavier@xmpuspus.dev" }]
keywords = ["rag", "knowledge-graph", "benchmark", "neo4j", "chromadb", "llm", "retrieval", "evaluation", "vector-search", "rag-evaluation", "document-retrieval"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer==0.12.5",
    "click>=8.1,<8.2",
    "rich==13.9.4",
    "pydantic==2.10.4",
    "pydantic-settings==2.7.1",
    "anthropic==0.42.0",
    "neo4j==5.27.0",
    "chromadb==0.5.23",
    "jellyfish==1.1.3",
    "networkx==3.4.2",
    "beautifulsoup4==4.12.3",
    "pyyaml==6.0.2",
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "sse-starlette==2.2.1",
    "openai==2.24.0",
]

[project.optional-dependencies]
dev = ["pytest==8.3.4", "ruff==0.8.6", "pytest-asyncio==0.24.0"]
pdf = ["PyMuPDF==1.25.3"]
docx = ["mammoth==1.8.0"]
web = ["httpx==0.28.1"]
all-formats = ["PyMuPDF==1.25.3", "mammoth==1.8.0", "httpx==0.28.1"]

[project.scripts]
kb-arena = "kb_arena.cli:app"

[project.urls]
Homepage = "https://github.com/xmpuspus/kb-arena"
Documentation = "https://github.com/xmpuspus/kb-arena/tree/main/docs"
Repository = "https://github.com/xmpuspus/kb-arena"
Issues = "https://github.com/xmpuspus/kb-arena/issues"

[tool.hatch.build.targets.sdist]
exclude = ["docs/", "datasets/", "web/", "tests/", "cypher/", ".github/", "**/.gitkeep"]
include = ["kb_arena/", "results/aws-compute_*.json", "results/.gitkeep", "LICENSE", "README.md", "pyproject.toml"]

[tool.hatch.build.targets.wheel]
packages = ["kb_arena"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
